add convert.utils, reduce duplicated code
